refactor(SDK-6089): move shared five icons rendering into one base class
The small and big views were near copies. FiveIconContentView now owns
the text row, icon loading and click wiring, following the Vertical
Image pattern; the two views only pick a layout. TemplateRenderer no
longer needs casts to read the unloaded icon count, and a stale log
claiming the notification is suppressed is dropped, since the renderer
already logs the basic-template fallback.
